requirements.txt
flask
torch
transformers
opencv-python
easyocr
pillow
rapidfuzz
